requirements.txt
fastapi
pydantic
python-dotenv
sqlalchemy
uvicorn
python-jose
python-multipart
passlib[bcrypt]
psycopg2-binary
Docker